pyproject.toml
[tool.poetry]
name = "attestation"
version = "0.1.0"
description = ""
authors = ["kifusss <kifus_06@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
